* commit '190115094f23e44c95c67ff7a275b1b33cd67938':
  Upgrade backend and frontend infrastructures
  Reestablished a date field required by the INPI for the renewal xml file
  Forgotten in last commit -> the personal renewal.php file needs to be updated manually according to the example file
  Updated the renewal batch payment xml file for EPO, some cleaning of the renewal management code, and updated the install script
  Security update
  Infrastructure update
  Fixed some rule dates not showing (use_after, use_before) and reformatted some code
  Updated Laravel app files to v9
  Implemented access restrictions to resources not authorized to some users

# Conflicts:
#	composer.json
#	composer.lock
#	public/css/app.css
#	public/js/app.js
#	yarn.lock

+++ app/Http/Controllers/ActorController.php
@@ -15,6 +15,7 @@ class ActorController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Actor::class);
         $actor = new Actor;
         if ($request->filled('Name')) {
             $actor = $actor->where('name', 'like', $request->Name . '%');
@@ -42,6 +43,7 @@ class ActorController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Actor::class);
         $table = new Actor;
         //TODO getTableComments is the same as in Rule.php. To render common
         $actorComments = $table->getTableComments('actor');
@@ -56,6 +58,7 @@ class ActorController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Actor::class);
         $request->validate([
             'name' => 'required|max:100',
             'email' => 'email|nullable'
@@ -72,6 +75,7 @@ class ActorController extends Controller
      */
     public function show(Actor $actor)
     {
+        $this->authorize('view', $actor);
         $actorInfo = $actor->load(['company:id,name', 'parent:id,name', 'site:id,name', 'droleInfo', 'countryInfo:iso,name', 'country_mailingInfo:iso,name', 'country_billingInfo:iso,name', 'nationalityInfo:iso,name']);
         $actorComments = $actor->getTableComments('actor');
         return view('actor.show', compact('actorInfo', 'actorComments'));
@@ -97,6 +101,7 @@ class ActorController extends Controller
      */
     public function update(Request $request, Actor $actor)
     {
+        $this->authorize('update', $actor);
         $request->validate([
             'email' => 'email|nullable',
             'ren_discount' => 'numeric|min:0|max:1'
@@ -114,6 +119,7 @@ class ActorController extends Controller
      */
     public function destroy(Actor $actor)
     {
+        $this->authorize('delete', $actor);
         $actor->delete();
         return $actor;
     }
